pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
]
